# perf: reduce the lantency of popup toggles