--- a/tests/helper_tests.sh
+++ b/tests/helper_tests.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 set -euo pipefail
 
-CURRENT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
+CURRENT_DIR=$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)
 
 # shellcheck source=../src/helpers.sh
 source "$CURRENT_DIR/../src/helpers.sh"
@@ -12,8 +12,8 @@ test_fail() {
 }
 
 test_interpolate() {
-	format="$1" expected="$2"
-	result="$(interpolate "${@:3}" "$format")"
+	format=$1" expected="$2
+	result=$(interpolate "${@:3}" "$format")
 
 	if [[ $result != "$expected" ]]; then
 		test_fail "$result != $expected"