--- a/tests/helper_tests.sh
+++ b/tests/helper_tests.sh
@@ -12,21 +12,22 @@ test_fail() {
 }
 
 test_interpolate() {
-	format=$1" expected="$2
-	result=$(interpolate "${@:3}" "$format")
+	local result
+	result=$(interpolate "${@}" "$format")
 
 	if [[ $result != "$expected" ]]; then
 		test_fail "$result != $expected"
 	fi
 }
 
-test_interpolate \
-	"{session}/{project}/{popup_name}" \
-	"working/{project}/default" \
-	"session" "working" \
-	"popup_name" "default"
-test_interpolate \
-	"{var1}/{var2}/{var2}/{var1}" \
-	"value1/value2/value2/value1" \
-	"var1" "value1" \
-	"var2" "value2"
+format="{session}/{project}/{popup_name}"
+expected="working/{project}/default"
+test_interpolate session="working" popup_name="default"
+
+format="{var1}/{var2}/{var2}/{var1}"
+expected="value1/value2/value2/value1"
+test_interpolate var1="value1" var2="value2"
+
+format="{var1}/{var2}/{var2}/{var1}"
+expected="var1=value1/var2=value2/var2=value2/var1=value1"
+test_interpolate var1="var1=value1" var2="var2=value2"