Adopt some tricks to reduce the latency of popup toggles:

- [x] Fetch tmux options in batch
  - [x] Tests pass on Linux
  - [ ] Tests pass on macOS
- [ ] Reduce tmux command parsing calls

This PR reduces latency by ~50% on my tablet, which is relatively slow and lantency-sensitive compared to desktops.

Before:

```console
$ hyperfine "./src/toggle.sh -E --toggle-mode=force-open --on-init='detach'"
Benchmark 1: ./src/toggle.sh -E --toggle-mode=force-open --on-init='detach'
  Time (mean ± σ):     472.5 ms ±  29.3 ms    [User: 55.8 ms, System: 95.7 ms]
  Range (min … max):   426.6 ms … 517.0 ms    10 runs

```

After:

```console
$ hyperfine "./src/toggle.sh -E --toggle-mode=force-open --on-init='detach'"
Benchmark 1: ./src/toggle.sh -E --toggle-mode=force-open --on-init='detach'
  Time (mean ± σ):     287.0 ms ±  14.6 ms    [User: 23.1 ms, System: 39.1 ms]
  Range (min … max):   259.6 ms … 311.9 ms    10 runs
```